Move query and add validator.
