+++ Src/SSar.BC.WebCams.Core/Queries/GetCamImageQuery/GetCamImageQueryValidator.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using FluentValidation;
+
+namespace SSar.BC.WebCams.Core.Queries.GetCamImageQuery
+{
+    public class GetCamImageQueryValidator : AbstractValidator<GetCamImageQuery>
+    {
+        public GetCamImageQueryValidator()
+        {
+            RuleFor(r => r.GroupUrlName).NotEmpty();
+            RuleFor(r => r.CameraUrlName).NotEmpty();
+        }
+    }
+}
