+++ Src/SSar.Presentation.API/Controllers/WebCamsController.cs
@@ -2,6 +2,7 @@
 using System.Threading.Tasks;
 using MediatR;
 using Microsoft.AspNetCore.Mvc;
+using SSar.BC.WebCams.Core.Queries.GetCamImageQuery;
 using SSar.BC.WebCams.Core.ValueTypes;
 
 // using SSar.BC.WebCams.Core.Commands;
@@ -61,7 +62,7 @@ namespace SSar.Presentation.API.Controllers
         public async Task<IActionResult> GetImage([FromRoute] string groupName, [FromRoute] string cameraName)
         {
             var camImage = (await _mediator.Send(
-                new GetImageByCamNameQuery(groupName, cameraName)));
+                new GetCamImageQuery(groupName, cameraName)));
 
             return File(
                 camImage.ByteArray
